django_blog/requirements.txt
autopep8==1.5.3
boto3==1.14.6
botocore==1.17.6
dj-database-url==0.5.0
django-crispy-forms==1.9.1
django-environ==0.4.5
django-heroku==0.3.1
django-markdownx==3.0.1
django-multiselectfield==0.1.12
django-storages==1.9.1
django-taggit==1.3.0
email-validator==1.1.1
gunicorn==20.0.4
jmespath==0.10.0
Markdown==3.2.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
s3transfer==0.3.3
whitenoise==5.1.0
